up client

+++ commands/new_test.go
@@ -10,8 +10,8 @@ func TestCalcAmountUSDT(t *testing.T) {
 	amountUSDT := CalcAmountUSDT(200.32, "6")
 	fmt.Println(amountUSDT)
 
-	priceBTC := float32(98000.00)
-	availableUSDT := float32(10.00)
+	priceBTC := 98000.00
+	availableUSDT := 10.00
 
 	amountCycleBTC := CalcAmountBTC(availableUSDT, priceBTC)
 	fmt.Println(amountCycleBTC)
